quick restructure for maintainability
