+++ GPTM Cloud App/admin-app/src/app/authentication/auth.guard.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
-import { AuthService } from './auth.service';
+import { AuthService } from '../services/auth.service';
 import {Router} from '@angular/router';
 import { AngularFireAuth } from '@angular/fire/auth';
-import { UserService } from './user.service';
+import { UserService } from '../services/user.service';
 @Injectable()
 export class AuthGuard implements CanActivate {
