+++ GPTM Cloud App/admin-app/src/app/services/data.service.ts
@@ -1,0 +1,18 @@
+import { Injectable } from '@angular/core';
+import * as firebase from 'firebase';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class DataService {
+
+  constructor() { }
+
+  writeUserData(userId, email, golfCourse) {
+    firebase.database().ref('users/' + userId).set({
+      email: email,
+      isAdmin: true,
+      golfCourse: golfCourse
+    });
+  }
+}
